서비스 단 인터페이스 도입, 상품 저장 시 영속성 전이 활용
