+++ ecommerce/src/main/java/com/hello/ecommerce/service/ProductCategoryService.java
@@ -19,85 +19,6 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
 
-@Service
-@RequiredArgsConstructor
-public class ProductCategoryService {
-
-    private final ProductCategoriesRepository productCategoriesRepository;
-    private final CategoryService categoryService;
-
-    public void saveAll(List<CategoryDto> categoryDtoList, Products saved) {
-        List<ProductCategories> productCategories = categoryDtoList.stream()
-                .map(cg -> {
-                    Categories category = categoryService.selectById(cg.getCategoryId());
-                    return ProductCategories.builder()
-                            .category(category)
-                            .isPrimary(cg.getIsPrimary())
-                            .products(saved)
-                            .build();
-                }).toList();
-        productCategoriesRepository.saveAll(productCategories);
-    }
-
-    public ProductCategoryResDto findProductsByCategoryId(ProductCategoryReqDto dto) {
-        Pageable pageable = PageRequest.of(dto.getPage(), dto.getPerPage());
-        Page<ProductCategories> result = productCategoriesRepository.findByCategoryId(dto.getCategoryId(), pageable);
-
-        CategoryDto categoryDto = getCategoryDto(result);
-        List<ProductResDto> productResDtoList = getProductResDtos(dto, result);
-        PaginationDto paginationDto
-                = new PaginationDto(result.getTotalElements(), result.getTotalPages(), dto.getPage() + 1, dto.getPerPage());
-
-        return new ProductCategoryResDto(categoryDto, productResDtoList, paginationDto);
-
-    }
-
-    private List<ProductResDto> getProductResDtos(ProductCategoryReqDto dto, Page<ProductCategories> result) {
-        List<Products> productsList = result.getContent().stream().map(ProductCategories::getProducts).toList();
-        productsList = productsList.stream().sorted(getComparator(dto.getSort())).toList();
-        return productsList.stream().map(product -> {
-            List<ProductImages> productImages = product.getProductImages().stream().filter(ProductImages::getIsPrimary).toList();
-            return new ProductResDto().builder()
-                    .id(product.getId())
-                    .name(product.getName())
-                    .slug(product.getSlug())
-                    .shortDescription(product.getShortDescription())
-                    .basePrice(product.getPrices().getBasePrice())
-                    .salePrice(product.getPrices().getSalePrice())
-                    .currency(product.getPrices().getCurrency())
-                    .primaryImage(new ImageDto(productImages.get(0).getAltText(), productImages.get(0).getUrl()))
-                    .brand(new BrandDto(product.getBrand().getId(), product.getBrand().getName()))
-                    .rating(product.getSeller().getRating())
-                    .reviewCount(product.getReviewsList().size())
-                    .createdAt(product.getCreatedAt())
-                    .build();
-        }).toList();
-    }
-
-    private CategoryDto getCategoryDto(Page<ProductCategories> result) {
-        ProductCategories productCategories = result.getContent().get(0);
-        Categories category = productCategories.getCategory();
-        return CategoryDto.builder().categoryId(category.getId())
-                .name(category.getName())
-                .slug(category.getSlug())
-                .description(category.getDescription())
-                .level(category.getLevel())
-                .imageUrl(category.getImageUrl())
-                .parent(new CategoryDto(category.getParent().getSlug(), category.getParent().getName(), category.getParent().getId()))
-                .build();
-    }
-
-    private static Comparator<Products> getComparator(String sort) {
-        Comparator<Products> comparator;
-        String[] parts = sort.split(":");
-        String sortField = parts[0];
-        String desc = parts[1];
-        comparator = switch (sortField) {
-            case "createdAt" -> Comparator.comparing(Products::getCreatedAt);
-            case "name" -> Comparator.comparing(Products::getName);
-            case "price" -> Comparator.comparing(p -> p.getPrices().getBasePrice());
-            default -> Comparator.comparing(Products::getCreatedAt);
-        };
-        return "desc".equals(desc) ? comparator.reversed() : comparator;
-    }
+public interface ProductCategoryService {
+    ProductCategoryResDto findProductsByCategoryId(ProductCategoryReqDto dto);
 }
