+++ ecommerce/src/main/java/com/hello/ecommerce/service/ProductOptionService.java
@@ -11,48 +11,12 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-@Service
-@RequiredArgsConstructor
-@Transactional
-public class ProductOptionService {
+public interface ProductOptionService {
+    ProductOptions selectById(Long id);
 
-    private final ProductOptionsRepository productOptionsRepository;
-    private final ProductOptionGroupsRepository productOptionGroupsRepository;
+    OptionDto save(OptionDto dto);
 
-    public List<ProductOptions> saveAll(List<OptionDto> optionDtoList, ProductOptionGroups groups) {
-        List<ProductOptions> options = optionDtoList.stream().map(op -> op.toEntity(groups)).toList();
-        return productOptionsRepository.saveAll(options);
-    }
+    OptionDto update(OptionDto dto);
 
-    public ProductOptions selectById(Long id) {
-        return productOptionsRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("상품 옵션이 존재하지 않습니다."));
-    }
-
-    public OptionDto save(OptionDto dto) {
-        ProductOptionGroups optionGroup = productOptionGroupsRepository.findById(dto.getOptionGroupId())
-                .orElseThrow(() -> new RuntimeException("상품 옵션 그룹 정보가 없습니다."));
-
-        ProductOptions entity = dto.toEntity(optionGroup);
-        ProductOptions saved = productOptionsRepository.save(entity);
-
-        return new OptionDto(saved);
-    }
-
-    public OptionDto update(OptionDto dto) {
-        ProductOptions productOptions = productOptionsRepository.findById(dto.getId())
-                .orElseThrow(() -> new RuntimeException("상품 옵션 정보가 없습니다."));
-
-        productOptions.setName(dto.getName());
-        productOptions.setAdditionalPrice(dto.getAdditionalPrice());
-        productOptions.setSku(dto.getSku());
-        productOptions.setStock(dto.getStock());
-        productOptions.setDisplayOrder(dto.getDisplayOrder());
-
-        return new OptionDto(productOptions);
-    }
-
-    public void delete(Long id) {
-        productOptionsRepository.deleteById(id);
-    }
+    void delete(Long id);
 }
