+++ ecommerce/src/main/java/com/hello/ecommerce/service/ProductTagServiceImpl.java
@@ -9,5 +9,8 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-public interface ProductTagService {
+@Service
+@RequiredArgsConstructor
+public class ProductTagServiceImpl implements ProductTagService {
+
 }
