+++ ecommerce/src/main/java/com/hello/ecommerce/service/SellerService.java
@@ -5,13 +5,6 @@ import com.hello.ecommerce.repository.SellersRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-@Service
-@RequiredArgsConstructor
-public class SellerService {
-
-    private final SellersRepository sellersRepository;
-
-    public Sellers selectById(Long id) {
-        return sellersRepository.findById(id).orElseThrow(() -> new RuntimeException("존재하지 않는 셀러입니다."));
-    }
+public interface SellerService {
+    Sellers selectById(Long id);
 }
